fix: update Hermes dSYM path for Catalyst (#36126)
Summary:
After a [change](https://github.com/facebook/hermes/commit/4ff7c31fb43f534947cca591e1f156d8c45aef9c) that landed on Hermes, the build script for Hermes generated an invalid dSYM name for Catalyst builds, which make our CI fail.
This PR aims to patch it while the Hermes team works on a long term fix for it.

The CI will go back to red as soon as that happen, but in that case, we will just need to revert this change. But, at least, we can go back to a green state for CircleCI.

## Changelog

[Internal] - update Hermes build script to be more resilient from dSYM names

Pull Request resolved: https://github.com/facebook/react-native/pull/36126

Test Plan: CircleCI is green

Reviewed By: cortinico

Differential Revision: D43193705

Pulled By: cipolleschi

fbshipit-source-id: a349dde8556564dacf4852f050610ae59139ffd8