Ability to always log certain performance entry types (#36820)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/36820

## Changelog:
[Internal] -

A follow-up to the D44584166 (which was abandoned, but this change makes sense nevertheless).

This allows to selectively enable logging of certain event types regardless of whether they are observed or not. For now it's marks and measures, but potentially it may be also e.g. navigation/resource entries.

Also expands unit tests for the JS side of the `Performance` API correspondingly.

Note that "always logged" and "observed" have different semantics. An "always logged" entry won't be sent back from native to JS, unless either:
* explicitly requested via `Performance.getEntries*`
* actually observed via `PerformanceObserver`

Reviewed By: rubennorte

Differential Revision: D44712550

fbshipit-source-id: 7fc891b09bd00fa9b510d1dc059cf908d5caea07