Remove stale CODEOWNERS (#36199)
Summary:
The [code owners](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners) configuration for this repository is stale, as most of the code owners no longer have write access to this repository. To my understanding, the code owners definition is not taken into account in the open source workflow for this repository. Therefore, instead of assigning new code owners, I'm removing it in this PR.

> **Note:** This has the side effect of no longer assigning me as a reviewer for PRs on this repository that touch the filepaths covered here.

<img width="1212" alt="Screenshot 2023-02-17 at 9 33 56 AM" src="https://user-images.githubusercontent.com/165856/219713156-816c293b-5568-49b6-ac56-06334cb00ca2.png">

## Changelog

[INTERNAL] - Remove CODEOWNERS. This is a GitHub-only change.

Pull Request resolved: https://github.com/facebook/react-native/pull/36199

Test Plan: N/A

Reviewed By: rshest

Differential Revision: D45274456

Pulled By: cortinico

fbshipit-source-id: baaca77f0bc9f64f93fcd24ddb86424d3ad50a43