Make createConstantsForViewManager work correctly with read-only maps (#37132)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37132

Changelog: [Internal]

The root cause was that one of the custom view managers was returning a read-only map for getExportedCustomDirectEventTypeConstants. React Native merges these maps between view managers, so if the view managers were initialized in a particular order, it will try to merge events into the read-only map and silently crashes.

Reviewed By: rshest

Differential Revision: D45370622

fbshipit-source-id: 7b4b4de372844835d60f81b6700438f56a6b9302